problems 1 and 2 plus build system
The solution for problem N will be in solutions/N.cpp. The
solutions/gen* scripts generate the int->solution map in map.cpp by
looking at which N.cpp files exist.

euler.cpp is the main function, whose only job is to read an integer and
then look it up in the map.